requirements.txt
accelerate==1.0.1
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.1.1
cycler==0.12.1
defusedxml==0.7.1
diffusers==0.30.3
distro==1.9.0
exceptiongroup==1.2.2
fairscale==0.4.4
filelock==3.13.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.3
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@126abe633ffe333e16e4a0a4e946bc1003caf757#egg=groundingdino&subdirectory=GroundingDINO
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.3
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
litellm==1.49.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.0
numpy==1.23.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
openai==1.51.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
propcache==0.2.0
psutil==6.1.1
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
-e git+https://github.com/xinyu1205/recognize-anything.git@88c2b0ca13e38cca6655f83ad0185271167dbcbf#egg=ram
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
safetensors==0.4.5
scipy==1.10.0
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@126abe633ffe333e16e4a0a4e946bc1003caf757#egg=segment_anything&subdirectory=segment_anything
six==1.16.0
sniffio==1.3.1
supervision==0.24.0
sympy==1.12
tenacity==9.0.0
tiktoken==0.7.0
timm==0.4.12
tokenizers==0.20.1
tomli==2.0.2
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.5
transformers==4.45.2
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yapf==0.40.2
yarl==1.14.0
zipp==3.20.2
